others/face_recognition/real_time_recognition/requirements.txt
cv2
numpy
dlib==19.8.1
face_recognition==1.2.2
pillow
